fraud_detection_common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fraud_detection_common"
version = "0.1.0"
description = "Common utilities for fraud detection"
requires-python = ">=3.9"
dependencies = [
    "sqlalchemy",
    "psycopg2-binary",
    "numpy",
    "pydantic",
    "pgvector",
    "python-dotenv",
    "scikit-learn"
]

[tool.hatch.build.targets.wheel]
packages = ["src/fraud_detection_common"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 